tests/test_game/test_game_logic.py
import unittest
from game import GameLogic

class TestGameLogic(unittest.TestCase):
    def test_check_win(self):
        game_logic = GameLogic()
        board = [...]
        win = game_logic.check_win(board)
        self.assertIsNotNone(win)

    def test_check_loss(self):
        game_logic = GameLogic()
        board = [...]
        loss = game_logic.check_loss(board)
        self.assertIsNotNone(loss)

    def test_check_draw(self):
        game_logic = GameLogic()
        board = [...]
        draw = game_logic.check_draw(board)
        self.assertIsNotNone(draw)

if __name__ == '__main__':
    unittest.main()
